# src/main/java/br/gov/caixa/api/investimentos/resource/produto/ProdutoResource.java
package br.gov.caixa.api.investimentos.resource.produto;

import jakarta.annotation.security.RolesAllowed;
import jakarta.inject.Inject;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;
import jakarta.ws.rs.*;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;
import br.gov.caixa.api.investimentos.dto.produto.ProdutoRequest;
import br.gov.caixa.api.investimentos.dto.produto.ProdutoResponse;
import br.gov.caixa.api.investimentos.exception.produto.ProdutoNotFoundException;
import br.gov.caixa.api.investimentos.enums.produto.TipoProduto;
import br.gov.caixa.api.investimentos.enums.produto.TipoRentabilidade;

import br.gov.caixa.api.investimentos.service.produto.ProdutoService;

import java.util.List;
import java.util.Optional;

@Path("/produtos")
@Produces(MediaType.APPLICATION_JSON)
@Consumes(MediaType.APPLICATION_JSON)
public class ProdutoResource {

    @Inject
    ProdutoService produtoService;

    /**
     * GET /produtos - Lista todos os produtos
     * Suporta filtros opcionais via query parameters
     */
    @GET
    @RolesAllowed({"USER", "ADMIN"})
    public Response listarProdutos(
            @QueryParam("tipo") TipoProduto tipo,
            @QueryParam("tipo_rentabilidade") TipoRentabilidade tipoRentabilidade,
            @QueryParam("fgc") Boolean fgc,
            @QueryParam("liquidez_imediata") Boolean liquidezImediata,
            @QueryParam("sem_liquidez") Boolean semLiquidez,
            @QueryParam("nome") String nome
    ) {
        List<ProdutoResponse> produtos;

        // Aplicar filtros baseados nos query parameters
        if (tipo != null) {
            produtos = produtoService.buscarPorTipo(tipo);
        } else if (tipoRentabilidade != null) {
            produtos = produtoService.buscarPorTipoRentabilidade(tipoRentabilidade);
        } else if (Boolean.TRUE.equals(fgc)) {
            produtos = produtoService.buscarProdutosComFgc();
        } else if (Boolean.TRUE.equals(liquidezImediata)) {
            produtos = produtoService.buscarProdutosComLiquidezImediata();
        } else if (Boolean.TRUE.equals(semLiquidez)) {
            produtos = produtoService.buscarProdutosSemLiquidez();
        } else if (nome != null && !nome.trim().isEmpty()) {
            produtos = produtoService.buscarPorNome(nome);
        } else {
            produtos = produtoService.listarTodos();
        }

        return Response.ok(produtos).build();
    }

    /**
     * GET /produtos/{id} - Busca produto por ID
     */
    @GET
    @Path("/{id}")
    @RolesAllowed({"USER", "ADMIN"})
    public Response buscarPorId(@PathParam("id") @NotNull Long id) {
        Optional<ProdutoResponse> produto = produtoService.buscarPorId(id);
        
        if (produto.isPresent()) {
            return Response.ok(produto.get()).build();
        } else {
            throw new ProdutoNotFoundException("Produto não encontrado com ID: " + id);
        }
    }

    /**
     * POST /produtos - Cria um novo produto
     */
    @POST
    @RolesAllowed({"ADMIN"})
    public Response criarProduto(@Valid @NotNull ProdutoRequest request)
    {
        ProdutoResponse produto = produtoService.criar(request);
        return Response.status(Response.Status.CREATED)
                .entity(produto)
                .build();
    }

    /**
     * PUT /produtos/{id} - Atualiza um produto existente
     */
    @PUT
    @Path("/{id}")
    @RolesAllowed({"ADMIN"})
    public Response atualizarProduto(@PathParam("id") @NotNull Long id, @Valid @NotNull ProdutoRequest request) {
        ProdutoResponse produto = produtoService.atualizar(id, request);
        return Response.ok(produto).build();
    }

    /**
     * GET /produtos/count - Conta total de produtos
     */
    @GET
    @Path("/count")
    @RolesAllowed({"USER", "ADMIN"})
    public Response contarProdutos() {
        long total = produtoService.contarTodos();
        return Response.ok(new CountResponse(total)).build();
    }

    /**
     * Classe para resposta de contagem
     */
    public static class CountResponse {
        private long total;

        public CountResponse() {}

        public CountResponse(long total) {
            this.total = total;
        }

        public long getTotal() {
            return total;
        }

        public void setTotal(long total) {
            this.total = total;
        }
    }
}

