# swagger-api-investimentos-caixa.yaml
openapi: 3.0.3
info:
  title: API Investimentos CAIXA
  description: |
    API RESTful para simulação e gestão de investimentos da CAIXA.
    
    Esta API permite:
    - Autenticação de usuários (clientes e administradores)
    - Gestão de clientes e perfis de risco
    - Catálogo completo de produtos de investimento
    - Simulação de investimentos com machine learning
    - Registros de investimentos reais
    - Recomendação personalizada de produtos
    - Telemetria e monitoramento do sistema
    
    ## Autenticação
    A API utiliza JWT (JSON Web Token) para autenticação. Use o endpoint `/entrar` para obter o token
    e inclua-o no header `Authorization: Bearer <token>` nas requisições autenticadas.
    
    ## Perfis de Usuário
    - **USER**: Clientes regulares (acesso limitado aos próprios dados)
    - **ADMIN**: Administradores (acesso completo ao sistema)
  version: 1.0.0
  contact:
    name: Equipe API Investimentos CAIXA
    email: api-investimentos@caixa.gov.br
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Ambiente de desenvolvimento
  - url: https://api-investimentos.caixa.gov.br
    description: Ambiente de produção
security:
  - bearerAuth: []
paths:
...
  /produtos:
    get:
      tags:
        - Produtos
      summary: Listar produtos
      description: |
        Lista produtos de investimento com filtros opcionais.
        Suporta múltiplos filtros via query parameters para refinar a busca.
      parameters:
        - name: tipo
          in: query
          description: Filtrar por tipo de produto
          schema:
            $ref: '#/components/schemas/TipoProduto'
        - name: tipo_rentabilidade
          in: query
          description: Filtrar por tipo de rentabilidade
          schema:
            $ref: '#/components/schemas/TipoRentabilidade'
        - name: fgc
          in: query
          description: Filtrar produtos com garantia FGC
          schema:
            type: boolean
        - name: liquidez_imediata
          in: query
          description: Filtrar produtos com liquidez imediata (0 dias)
          schema:
            type: boolean
        - name: sem_liquidez
          in: query
          description: Filtrar produtos sem liquidez (-1)
          schema:
            type: boolean
        - name: nome
          in: query
          description: Buscar produtos por nome (busca parcial)
          schema:
            type: string
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProdutoResponse'
    
    post:
      tags:
        - Produtos
      summary: Criar novo produto
      description: Cria um novo produto de investimento (apenas para ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProdutoRequest'
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /produtos/{id}:
    get:
      tags:
        - Produtos
      summary: Buscar produto por ID
      description: Retorna detalhes de um produto específico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detalhes do produto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoResponse'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - Produtos
      summary: Atualizar produto
      description: Atualiza um produto existente (apenas para ADMIN)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProdutoRequest'
      responses:
        '200':
          description: Produto atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoResponse'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /telemetria:
    get:
      tags:
        - Telemetria
      summary: Obter telemetria do sistema
      description: Retorna métricas consolidadas do sistema (apenas para ADMIN)
      responses:
        '200':
          description: Métricas do sistema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetriaResponse'
        '500':
          description: Erro ao obter telemetria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Telemetria
      summary: Limpar métricas
      description: Limpa todas as métricas armazenadas (apenas para ADMIN)
      responses:
        '204':
          description: Métricas limpas com sucesso
        '500':
          description: Erro ao limpar métricas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /telemetria/mais-acessados/{limite}:
    get:
      tags:
        - Telemetria
      summary: Endpoints mais acessados
      description: Retorna os endpoints mais acessados (apenas para ADMIN)
      parameters:
        - name: limite
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100
            example: 10
      responses:
        '200':
          description: Endpoints mais acessados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetriaMetrica'

  # TELEMETRIA - ACESSO LOGS
  /telemetria/acesso-logs:
    get:
      tags:
        - Telemetria
      summary: Lista todos os logs de acesso
      description: Retorna todos os registros de acesso (apenas ADMIN). Recomendado usar filtros/paginação.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            example: 0
        - name: size
          in: query
          schema:
            type: integer
            example: 50
      responses:
        '200':
          description: Lista de logs de acesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AcessoLog'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Telemetria
      summary: Limpar TODOS os logs de acesso
      description: Operação destrutiva que remove todo o histórico de logs (apenas ADMIN)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logs removidos
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /telemetria/acesso-logs/usuario/{usuarioId}:
    get:
      tags:
        - Telemetria
      summary: Buscar logs por usuário
      parameters:
        - name: usuarioId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logs do usuário
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AcessoLog'
  /telemetria/acesso-logs/erros:
    get:
      tags:
        - Telemetria
      summary: Buscar logs com erro (status >= 400)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logs com erro
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AcessoLog'
  /telemetria/acesso-logs/status/{statusCode}:
    get:
      tags:
        - Telemetria
      summary: Buscar logs por status code
      parameters:
        - name: statusCode
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logs filtrados por status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AcessoLog'
  /telemetria/acesso-logs/estatisticas:
    get:
      tags:
        - Telemetria
      summary: Estatísticas dos logs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Estatísticas de acesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstatisticasAcesso'
  /telemetria/acesso-logs/antigos/{diasRetencao}:
    delete:
      tags:
        - Telemetria
      summary: Limpar logs antigos
      parameters:
        - name: diasRetencao
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logs antigos removidos
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtido através do endpoint /entrar
  schemas:
...
    TipoProduto:
      type: string
      enum: [CDB, LCI, LCA, POUPANCA, TESOURO_DIRETO, DEBENTURE, CRI, FUNDO, FII, ACAO, ETF]
      example: "CDB"
    TipoRentabilidade:
      type: string
      enum: [PRE, POS]
      example: "PRE"
    PeriodoRentabilidade:
      type: string
      enum: [AO_DIA, AO_MES, AO_ANO, PERIODO_TOTAL]
      example: "AO_ANO"
    Indice:
      type: string
      enum: [NENHUM, CDI, SELIC, IPCA, IGP_M, IGP_DI, INCC, IPC_A]
      example: "CDI"
    NivelRisco:
      type: string
      enum: [BAIXO, MEDIO, ALTO]
      example: "BAIXO"
    ProdutoRequest:
      type: object
      required: [nome, tipo, tipo_rentabilidade, rentabilidade, periodo_rentabilidade, liquidez, minimo_dias_investimento, fgc]
      properties:
        nome:
          type: string
          minLength: 2
          maxLength: 255
          example: "CDB Caixa Progressivo"
        tipo:
          $ref: '#/components/schemas/TipoProduto'
        tipo_rentabilidade:
          $ref: '#/components/schemas/TipoRentabilidade'
        rentabilidade:
          type: number
          format: decimal
          minimum: 0
          example: 10.5
          description: Taxa de rentabilidade (percentual)
        periodo_rentabilidade:
          $ref: '#/components/schemas/PeriodoRentabilidade'
        indice:
          $ref: '#/components/schemas/Indice'
        liquidez:
          type: integer
          minimum: -1
          example: 90
          description: Dias para liquidez (-1 = sem liquidez, 0 = liquidez imediata)
        minimo_dias_investimento:
          type: integer
          minimum: 0
          example: 90
          description: Período mínimo de investimento em dias
        fgc:
          type: boolean
          example: true
          description: Possui garantia do Fundo Garantidor de Créditos
    ProdutoResponse:
      allOf:
        - $ref: '#/components/schemas/ProdutoRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              example: 1
            risco:
              $ref: '#/components/schemas/NivelRisco'

    # INVESTIMENTOS
...
    TelemetriaResponse:
      type: object
      properties:
        periodoPesquisa:
          type: string
          example: "2024-01-01T00:00:00 até 2024-01-15T23:59:59"
        totalEndpoints:
          type: integer
          example: 8
        metricas:
          type: array
          items:
            type: object
            properties:
              endpoint:
                type: string
                example: "/produtos"
              execucoes:
                type: integer
                example: 150
              tempoMedio:
                type: number
                format: decimal
                example: 125.5
                description: Tempo médio de resposta em millisegundos
    TelemetriaMetrica:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        endpoint:
          type: string
          example: "/produtos"
        contador:
          type: integer
          example: 150
        tempoMedio:
          type: number
          format: decimal
          example: 125.5
          description: Tempo médio em millisegundos
        ultimaAtualizacao:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"

    # ERRO
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Credenciais inválidas"
          description: Descrição do erro
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
        status:
          type: integer
          example: 401
          description: Código de status HTTP
        path:
          type: string
          example: "/entrar"
          description: Path da requisição que gerou o erro
        details:
          type: object
          nullable: true
          description: Detalhes adicionais do erro
    AcessoLog:
      type: object
      properties:
        id:
          type: integer
          format: int64
        usuarioId:
          type: integer
          format: int64
          nullable: true
        usuarioNome:
          type: string
          nullable: true
        endpoint:
          type: string
        metodoHttp:
          type: string
        uriCompleta:
          type: string
        ipOrigem:
          type: string
        corpoRequisicao:
          type: string
          description: "Truncado a 10KB"
        statusCode:
          type: integer
        corpoResposta:
          type: string
          description: "Truncado a 10KB"
        tempoExecucaoMs:
          type: integer
        dataAcesso:
          type: string
          format: date-time
        userAgent:
          type: string
        erroMessage:
          type: string
          nullable: true
        erroStacktrace:
          type: string
          nullable: true
    EstatisticasAcesso:
      type: object
      properties:
        totalAcessos:
          type: integer
          format: int64
        acessosComSucesso:
          type: integer
          format: int64
        acessosComErro:
          type: integer
          format: int64
        taxaSucesso:
          type: number
          format: float
        taxaErro:
          type: number
          format: float
...